Fixes to make codeclimate happy
